some brief changes
