+++ lib/operations/yeah.js
@@ -7,7 +7,7 @@ module.exports = {
 
   register (app, output) {
     return app.get('/yeah/:from', function (req, res) {
-      const message = 'fork YEAH!'
+      const message = 'Yeah, you do it your way and I will do it the correct way.'
       const subtitle = `- ${req.params.from}`
       return output(req, res, message, subtitle)
     })
