+++ spec/operations/desire_spec.js
@@ -1,0 +1,44 @@
+const operation = require('../../lib/operations/desire')
+
+describe('/desire', function () {
+  it('should have the correct name', () => expect(operation.name).toEqual('Desire'))
+
+  it('should have the correct url', () => expect(operation.url).toEqual('/desire/:from'))
+
+  it('should have the correct fields', () =>
+    expect(operation.fields).toEqual([
+      { name: 'From', field: 'from' }
+    ])
+  )
+
+  return describe('register', function () {
+    it('should call app.get with correct url', function () {
+      const app =
+        { get: jasmine.createSpy() }
+
+      operation.register(app, null)
+
+      expect(app.get).toHaveBeenCalledWith('/desire/:from', jasmine.any(Function))
+    })
+
+    return it('should call output with correct params', function () {
+      let func = null
+      const app =
+        { get (url, fn) { return func = fn } }
+      const output = jasmine.createSpy()
+      operation.register(app, output)
+
+      const req = {
+        params: {
+          from: 'TESTFROM'
+        }
+      }
+
+      const message = 'I desire that we be better strangers.'
+      const subtitle = `- ${req.params.from}`
+
+      func(req, 'RES')
+      return expect(output).toHaveBeenCalledWith(req, 'RES', message, subtitle)
+    })
+  })
+})
